% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{fars_map_state}
\alias{fars_map_state}
\title{plot a map of the data in states}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{the number of the state}

\item{year}{year we want to plot}
}
\value{
nothing but a plot will be generated
}
\description{
This function plot a map of data in given states in the given year
}
\examples{
fars_map_state(1, 2013)
}
