% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read data file and return it}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{the file name of the data file}
}
\value{
a data.frame.
}
\description{
This function reads the data file 'filename' and return
it as a data.frame.
}
