% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read the data with a specified year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{years}
}
\value{
data with specified years and months. NULL if error is raised
}
\description{
This function will first create a file name using make_filename containing the
given years, and then read data from those files. It will return the data with the specified years.
If those years cannot be found or are not valid, it raise an error.
}
\examples{
fars_read_years(2013)
}
