% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize data by years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{years}
}
\value{
summarized data by years and count
}
\description{
This function will read data from the file name that's created by fars_read_years with the provided years.
After that, it will summarize the data by the years
}
\examples{
fars_summarize_years(2013)
}
