% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a file name containing the year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{year name}
}
\value{
a character string
}
\description{
This function creates a string containing the provided year, which will be used as the file name
}
\examples{
make_filename(2013)
}
